+++ completion/git-machete.completion.bash
@@ -22,7 +22,7 @@ _git_machete() {
   local discover_opts="-C --checked-out-since= -l --list-commits -r --roots= -y --yes"
   local fork_point_opts="--inferred --override-to= --override-to-inferred --override-to-parent --unset-override"
   local github_anno_prs_opts="--with-urls"
-  local github_create_pr_opts="--draft"
+  local github_create_pr_opts="--draft --title="
   local github_checkout_prs_opts="--all --by= --mine"
   local github_retarget_pr_opts="-b --branch= --ignore-if-missing"
   local reapply_opts="-f --fork-point="